Create JSONResponse trait
