+++ app/Http/Controllers/Auth/JSONResponse.php
@@ -1,0 +1,31 @@
+<?php
+
+/**
+ *	Generated by IceTea Framework 0.0.1
+ *	Created at 2017-12-14 13:00:50
+ *	Namespace App\Http\Controllers\Auth
+ */
+
+namespace App\Http\Controllers\Auth;
+
+use IceTea\Http\Controller;
+
+trait JSONResponse
+{
+	public function err($msg, $url = null)
+    {
+        http_response_code(400);
+        exit($this->buildJson(
+            [
+                "status"   => "error",
+                "message"  => $msg,
+                "redirect" => $url
+            ]
+        ));
+    }
+
+    private function buildJson($data)
+    {
+        return json_encode($data, JSON_UNESCAPED_SLASHES);
+    }
+}
